Standardize on "unsigned short" for server ports
The "server.h" file contains the prototype for the "server_new"
function. This function expects an "unsigned short" for port.
This patch makes the code consistently use "unsigned short",
rather than a mix of "unsigned short" and "int" types.

Even better, might be to use "uint16_t", however first it should
be demonstrated that this can be done without creating platform
and toolchain compatibility headaches.

See also: https://tools.ietf.org/html/rfc6335#section-6
And:
$ grep typedef -r /usr/include/netinet | grep port
/usr/include/netinet/in.h:typedef uint16_t in_port_t;

+++ conffile.tab.c
@@ -134,7 +134,7 @@ struct _clust {
 };
 struct _clhost {
 	char *ip;
-	int port;
+	unsigned short port;
 	char *inst;
 	int proto;
 	con_type type;
@@ -162,7 +162,7 @@ struct _lsnr {
 struct _rcptr {
 	con_proto ctype;
 	char *ip;
-	int port;
+	unsigned short port;
 	void *saddr;
 	struct _rcptr *next;
 };
